+++ packages/x-components/src/x-modules/ai/store/actions/fetch-and-save-ai-suggestions.action.ts
@@ -1,14 +1,14 @@
-import type { AiQuestion } from '@empathyco/x-types'
+import type { AiQuestion, AiSuggestionQuery } from '@empathyco/x-types'
 import type { AiActionContext, AiXStoreModule } from '../types'
 import { XPlugin } from '../../../../plugins'
 
 type AnswerChunk =
   | { responseText: string }
   | { suggestionText: string }
-  | { queries: { query: string; categories: string[] }[] }
+  | { queries: AiSuggestionQuery[] }
   | { taggings: AiQuestion['tagging'][] }
 /**
- * Default implementation for the {@link AiActions.fetchAiSuggestions}.
+ * Default implementation for the {@link AiActions.fetchAndSaveAiSuggestions}.
  *
  * @param _ - The {@link https://vuex.vuejs.org/guide/actions.html | context} of the actions,
  * provided by Vuex.
@@ -17,37 +17,22 @@ type AnswerChunk =
  *
  * @public
  */
-export const fetchAiSuggestions: AiXStoreModule['actions']['fetchAiSuggestions'] = async (
-  { commit },
-  request,
-) => {
-  if (!request) {
-    return null
+export const fetchAndSaveAiSuggestions: AiXStoreModule['actions']['fetchAndSaveAiSuggestions'] =
+  async ({ commit }, request) => {
+    if (!request) {
+      return
+    }
+    commit('setSuggestionsLoading', true)
+    return XPlugin.adapter.aiSuggestions(request).then(({ body, status }) => {
+      if (status !== 200) {
+        return
+      }
+      if (body) {
+        const reader = body.getReader()
+        readAnswer(reader, commit)
+      }
+    })
   }
-  commit('setSuggestionsLoading', true)
-  return (
-    XPlugin.adapter
-      /*.aiSuggestions({
-        query: 'toner',
-        extraParams: {
-          lang: 'es',
-          env: 'staging',
-          instance: 'mymotivemarketplace',
-          filters: { store: 'chupatintas-s-l', city: 'oviedo' },
-        },
-      })*/
-      .aiSuggestions(request)
-      .then(({ body, status }) => {
-        if (status !== 200) {
-          return null
-        }
-        if (body) {
-          const reader = body.getReader()
-          readAnswer(reader, commit)
-        }
-      })
-  )
-}
 
 function readAnswer(
   reader: ReadableStreamDefaultReader<Uint8Array>,
@@ -66,7 +51,8 @@ function readAnswer(
         let data: AnswerChunk
 
         for (const line of lines) {
-          if (line.startsWith('data:')) {
+          // line.length check to avoid empty data chunks
+          if (line.startsWith('data:') && line.length > 5) {
             data = JSON.parse(line.slice(5).trim()) as AnswerChunk
             if ('responseText' in data) {
               commit('setResponseText', data.responseText)