+++ packages/x-components/src/x-modules/ai/store/getters/index.ts
@@ -1,2 +1,3 @@
 export { query as aiQuery } from './ai-query.getter'
-export { request as aiQuestionsRequest } from './request.getter'
+export { suggestionsRequest as aiSuggestionsRequest } from './suggestions-request.getter'
+export { suggestionsSearchRequest as aiSuggestionsSearchRequest } from './suggestions-search-request.getter'