+++ packages/x-components/src/x-modules/ai/store/getters/suggestions-search-request.getter.ts
@@ -1,0 +1,18 @@
+import type { AiXStoreModule } from '../types'
+
+/**
+ * Default implementation for the {@link AiGetters.suggestionsSearchRequest} getter.
+ *
+ * @param state - Current {@link https://vuex.vuejs.org/guide/state.html | state} of the ai
+ * module.
+ *
+ * @returns The ai request to fetch data from the API.
+ *
+ * @public
+ */
+export const suggestionsSearchRequest: AiXStoreModule['getters']['suggestionsSearchRequest'] = ({
+  params,
+  queries,
+}) => {
+  return queries.length ? { queries, extraParams: params } : null
+}